refactor: Improve CSV generation and update related resources
The code changes improve the generation of the CSV file by excluding the
header row. This change ensures that the CSV file is properly formatted
for use in downstream processes. Additionally, the related resources
section has been updated to include new links and remove obsolete ones,
providing users with up-to-date references.

Related resources:
- https://github.com/mrmaxguns/wonderwordsmodule
-
https://github.com/taikuukaits/SimpleWordlists/blob/master/Wordlist-Adjectives-All.txt
- https://github.com/aaronbassett/Pass-phrase
- ... (13 more references)